Implement spin to fms color command with woof
